requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.30
alembic==1.13.1
python-dotenv==1.0.0
aiohttp==3.9.5
aiosqlite==0.20.0
